refactor: move core parts of printer to separate workspace
