+++ src/packages/core-parts/package.json
@@ -1,14 +1,13 @@
 {
   "private": true,
-  "name": "v2-plugin",
+  "name": "core-parts",
   "version": "0.0.0",
   "author": "Hyeonjong <nianelo4@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "core-parts": "workspace:*",
-    "prettier": "2.8.4"
+    "diff": "5.1.0"
   },
   "devDependencies": {
-    "@types/prettier": "2.7.3"
+    "@types/diff": "5.0.3"
   }
 }
