+++ src/packages/v3-plugin/package.json
@@ -5,11 +5,10 @@
   "author": "Hyeonjong <nianelo4@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "diff": "5.1.0",
+    "core-parts": "workspace:*",
     "prettier": "3.0.3"
   },
   "devDependencies": {
-    "@types/diff": "5.0.3",
     "@types/node": "20.2.5"
   }
 }
